refactor: extract filter methods from view store
