+++ src/web/common/components/ContextMenu/HideMenuItem.tsx
@@ -1,5 +1,5 @@
 import { Node } from "../../../topic/utils/graph";
-import { hideNode } from "../../../view/currentViewStore/store";
+import { hideNode } from "../../../view/currentViewStore/filter";
 import { CloseOnClickMenuItem } from "./CloseOnClickMenuItem";
 
 export const HideMenuItem = ({ node }: { node: Node }) => {
