+++ src/web/topic/components/Indicator/ForceShownIndicator.tsx
@@ -2,7 +2,10 @@ import { WbTwilight } from "@mui/icons-material";
 import { type ButtonProps } from "@mui/material";
 import { MouseEventHandler, memo, useCallback } from "react";
 
-import { stopForcingNodeToShow, useIsNodeForcedToShow } from "../../../view/currentViewStore/store";
+import {
+  stopForcingNodeToShow,
+  useIsNodeForcedToShow,
+} from "../../../view/currentViewStore/filter";
 import { Indicator } from "./Indicator";
 
 interface Props {
