+++ src/web/view/components/Filter/GeneralFilters.tsx
@@ -11,7 +11,8 @@ import { Switch } from "../../../common/components/Form/Switch";
 import { deepIsEqual } from "../../../common/store/utils";
 import { useAllNodes } from "../../../topic/store/nodeHooks";
 import { possibleScores } from "../../../topic/utils/graph";
-import { setGeneralFilter, useFormat, useGeneralFilter } from "../../currentViewStore/store";
+import { setGeneralFilter, useGeneralFilter } from "../../currentViewStore/filter";
+import { useFormat } from "../../currentViewStore/store";
 import { GeneralFilter, generalFilterSchema, scoredComparers } from "../../utils/generalFilter";
 import { ShowSecondaryNeighborsLabel } from "./ShowSecondaryNeighborsLabel";
