+++ src/web/view/currentViewStore/store.ts
@@ -1,21 +1,14 @@
 import mergeWith from "lodash/mergeWith";
-import union from "lodash/union";
-import without from "lodash/without";
 import { temporal } from "zundo";
 import { useStore } from "zustand";
 import { devtools, persist } from "zustand/middleware";
-import { shallow } from "zustand/shallow";
 import { createWithEqualityFn } from "zustand/traditional";
 
 import { Format, InfoCategory } from "../../../common/infoCategory";
-import { infoNodeTypes, nodeTypes } from "../../../common/node";
+import { nodeTypes } from "../../../common/node";
 import { emitter } from "../../common/event";
 import { useGraphPart } from "../../topic/store/graphPartHooks";
-import { getDefaultNode } from "../../topic/store/nodeGetters";
-import { useTopicStore } from "../../topic/store/store";
-import { findNodeOrThrow } from "../../topic/utils/graph";
-import { neighbors } from "../../topic/utils/node";
-import { DiagramFilter, StandardFilter, StandardFilterWithFallbacks } from "../utils/diagramFilter";
+import { StandardFilter } from "../utils/diagramFilter";
 import { GeneralFilter } from "../utils/generalFilter";
 import { TableFilter } from "../utils/tableFilter";
 
@@ -118,66 +111,6 @@ export const useFormat = () => {
   return useCurrentViewStore((state) => state.format);
 };
 
-const useCategoriesToShow = () => {
-  return useCurrentViewStore((state) => state.categoriesToShow, shallow);
-};
-
-const useStandardFilter = (category: InfoCategory) => {
-  return useCurrentViewStore((state) => state[`${category}Filter`], shallow);
-};
-
-export const useDiagramFilter = (): DiagramFilter => {
-  const categoriesToShow = useCategoriesToShow();
-  const structureFilter = useStandardFilter("structure");
-  const researchFilter = useStandardFilter("research");
-  const justificationFilter = useStandardFilter("justification");
-
-  return {
-    structure: { show: categoriesToShow.includes("structure"), ...structureFilter },
-    research: { show: categoriesToShow.includes("research"), ...researchFilter },
-    justification: { show: categoriesToShow.includes("justification"), ...justificationFilter },
-  };
-};
-
-export const useTableFilter = (): TableFilter => {
-  return useCurrentViewStore(
-    (state) => state.tableFilter,
-    (before, after) => JSON.stringify(before) === JSON.stringify(after)
-  );
-};
-
-export const useTableFilterWithFallbacks = (): TableFilter => {
-  const tableFilter = useTableFilter();
-
-  const centralProblemId = getDefaultNode("problem")?.id;
-
-  const tableFilterDefaults = {
-    centralProblemId,
-    solutions: [],
-    criteria: [],
-  };
-
-  return { ...tableFilterDefaults, ...tableFilter };
-};
-
-export const useGeneralFilter = () => {
-  return useCurrentViewStore((state) => state.generalFilter);
-};
-
-export const useIsNodeForcedToShow = (nodeId: string) => {
-  return useCurrentViewStore((state) => state.generalFilter.nodesToShow.includes(nodeId));
-};
-
-export const usePrimaryNodeTypes = () => {
-  return useCurrentViewStore((state) => {
-    return state.categoriesToShow.flatMap((category) => infoNodeTypes[category]);
-  });
-};
-
-export const useShowImpliedEdges = () => {
-  return useCurrentViewStore((state) => state.showImpliedEdges);
-};
-
 export const useCanGoBackForward = () => {
   const temporalStore = useTemporalStore();
 
@@ -193,143 +126,6 @@ export const setSelected = (graphPartId: string | null) => {
 
 export const setFormat = (format: Format) => {
   useCurrentViewStore.setState({ format }, false, "setFormat");
-};
-
-export const setShowInformation = (category: InfoCategory, show: boolean) => {
-  const oldCategoriesToShow = useCurrentViewStore.getState().categoriesToShow;
-
-  const newCategoriesToShow =
-    show && !oldCategoriesToShow.includes(category)
-      ? [...oldCategoriesToShow, category]
-      : !show && oldCategoriesToShow.includes(category)
-      ? oldCategoriesToShow.filter((c) => c !== category)
-      : null;
-
-  if (newCategoriesToShow) {
-    useCurrentViewStore.setState(
-      { categoriesToShow: newCategoriesToShow },
-      false,
-      "setShowInformation"
-    );
-    emitter.emit("changedDiagramFilter");
-  }
-};
-
-export const setStandardFilter = (category: InfoCategory, filter: StandardFilter) => {
-  const newFilter =
-    category === "structure"
-      ? { structureFilter: filter }
-      : category === "research"
-      ? { researchFilter: filter }
-      : { justificationFilter: filter };
-
-  useCurrentViewStore.setState(newFilter, false, "setStandardFilter");
-
-  emitter.emit("changedDiagramFilter");
-};
-
-export const setTableFilter = (tableFilter: TableFilter) => {
-  useCurrentViewStore.setState({ tableFilter }, false, "setTableFilter");
-};
-
-export const setGeneralFilter = (generalFilter: GeneralFilter) => {
-  useCurrentViewStore.setState({ generalFilter }, false, "setGeneralFilter");
-  emitter.emit("changedDiagramFilter");
-};
-
-export const viewCriteriaTable = (problemNodeId: string) => {
-  useCurrentViewStore.setState(
-    {
-      format: "table",
-      tableFilter: {
-        centralProblemId: problemNodeId,
-        solutions: [],
-        criteria: [],
-      },
-    },
-    false,
-    "viewCriteriaTable"
-  );
-};
-
-export const viewJustification = (arguedDiagramPartId: string) => {
-  useCurrentViewStore.setState(
-    {
-      format: "diagram",
-      categoriesToShow: ["justification"],
-      justificationFilter: { type: "rootClaim", centralRootClaimId: arguedDiagramPartId },
-    },
-    false,
-    "viewJustification"
-  );
-};
-
-/**
- * @param also true if these nodes should be added to the current filter, false if they should be the only nodes displayed
- */
-export const showNodeAndNeighbors = (nodeId: string, also: boolean) => {
-  const { categoriesToShow, generalFilter } = useCurrentViewStore.getState();
-  const graph = useTopicStore.getState();
-  const node = findNodeOrThrow(nodeId, graph.nodes);
-
-  const nodeAndNeighbors = [nodeId, ...neighbors(node, graph).map((neighbor) => neighbor.id)];
-
-  useCurrentViewStore.setState({
-    format: "diagram",
-    categoriesToShow: also ? categoriesToShow : [],
-    generalFilter: {
-      ...generalFilter,
-      nodesToShow: also ? union(generalFilter.nodesToShow, nodeAndNeighbors) : nodeAndNeighbors,
-      nodesToHide: without(generalFilter.nodesToHide, ...nodeAndNeighbors),
-    },
-  });
-
-  emitter.emit("changedDiagramFilter");
-};
-
-export const stopForcingNodeToShow = (nodeId: string) => {
-  const generalFilter = useCurrentViewStore.getState().generalFilter;
-
-  useCurrentViewStore.setState({
-    generalFilter: {
-      ...generalFilter,
-      nodesToShow: without(generalFilter.nodesToShow, nodeId),
-    },
-  });
-};
-
-export const showNode = (nodeId: string) => {
-  const generalFilter = useCurrentViewStore.getState().generalFilter;
-
-  useCurrentViewStore.setState({
-    generalFilter: {
-      ...generalFilter,
-      nodesToShow: union(generalFilter.nodesToShow, [nodeId]),
-      nodesToHide: without(generalFilter.nodesToHide, nodeId),
-    },
-  });
-
-  emitter.emit("changedDiagramFilter");
-};
-
-export const hideNode = (nodeId: string) => {
-  const generalFilter = useCurrentViewStore.getState().generalFilter;
-
-  useCurrentViewStore.setState({
-    generalFilter: {
-      ...generalFilter,
-      nodesToShow: without(generalFilter.nodesToShow, nodeId),
-      nodesToHide: union(generalFilter.nodesToHide, [nodeId]),
-    },
-  });
-
-  emitter.emit("changedDiagramFilter");
-};
-
-export const toggleShowImpliedEdges = (show: boolean) => {
-  useCurrentViewStore.setState({ showImpliedEdges: show });
-
-  emitter.emit("changedDiagramFilter");
 };
 
 export const goBack = () => {
@@ -375,29 +171,4 @@ export const loadView = async (persistId: string) => {
 
   // it doesn't make sense to want to undo a page load
   useCurrentViewStore.temporal.getState().clear();
-};
-
-// helpers
-export const getStandardFilterWithFallbacks = (standardFilter: StandardFilter): StandardFilter => {
-  const centralProblemId = getDefaultNode("problem")?.id;
-  const centralSolutionId = getDefaultNode("solution")?.id;
-  const centralQuestionId = getDefaultNode("question")?.id;
-  const centralSourceId = getDefaultNode("source")?.id;
-  const centralRootClaimId = getDefaultNode("rootClaim")?.id;
-
-  const standardFilterDefaults: StandardFilterWithFallbacks = {
-    type: "none",
-    centralProblemId,
-    problemDetails: ["causes", "effects", "subproblems", "criteria", "solutions"],
-    centralSolutionId,
-    solutionDetail: "all",
-    solutions: [],
-    criteria: [],
-    centralQuestionId,
-    centralSourceId,
-    centralRootClaimId,
-  };
-
-  // override any defaults using the stored filter
-  return { ...standardFilterDefaults, ...standardFilter };
 };
